src/main/java/app/feature/controller/UserController.java
package app.feature.controller;

import app.core.apiversion.ApiVersion;
import app.feature.domain.User;
import app.feature.service.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping(value = "/users")
public class UserController {

    @Autowired
    private UserService userService;

    @GetMapping
    public List<User> getUsers(){
        return userService.getUsers();
    }


}
